requirements.txt
absl-py==0.9.0
addict==2.2.1
appdirs==1.4.3
astunparse==1.6.3
attrs==19.3.0
backcall==0.2.0
bleach==3.1.5
CacheControl==0.12.6
cachetools==4.1.1
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
configparser==5.0.0
contextlib2==0.6.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
docker-pycreds==0.4.0
dominate==2.5.1
entrypoints==0.3
future==0.18.2
gast==0.3.3
gitdb==4.0.5
GitPython==3.1.7
google-auth==1.18.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
gql==0.2.0
graphql-core==1.1
graphviz==0.14
grpcio==1.30.0
h5py==2.10.0
html5lib==1.0.1
idna==2.8
imageio==2.9.0
ipaddr==2.2.0
ipykernel==5.3.0
ipyplot==1.0.5
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.1
Jinja2==2.11.2
joblib==0.16.0
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lockfile==0.12.2
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.2
mistune==0.8.4
mmcv==1.0.5
mmfashion==0.4.0
msgpack==0.6.2
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.4
notebook==6.0.3
numpy==1.19.0
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
opencv-python==4.2.0.34
opt-einsum==3.2.1
packaging==20.3
pandas==1.0.5
pandocfilters==1.4.2
parso==0.7.0
pathtools==0.1.2
pep517==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
progress==1.5
prometheus-client==0.8.0
promise==2.3
prompt-toolkit==3.0.5
protobuf==3.12.2
psutil==5.7.2
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
pyparsing==2.4.6
pyrsistent==0.16.0
python-dateutil==2.8.1
pytoml==0.1.21
pytorch-metric-learning==0.9.88
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.1
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.4.1
Send2Trash==1.5.0
sentry-sdk==0.16.3
shortuuid==1.0.1
six==1.14.0
smmap==3.0.4
subprocess32==3.5.4
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
tensorboardX==2.1
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.7.4
torch==1.5.1
torchsummaryX==1.3.0
torchvision==0.6.1
torchviz==0.0.1
tornado==6.0.4
tqdm==4.47.0
traitlets==4.3.3
urllib3==1.25.8
wandb==0.9.4
watchdog==0.10.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
yapf==0.30.0
